PEEPFT-13 phytec_eeprom_flashtool.py: generalize sysfs EEPROM filepath
This commit forces the script to convert PHYTEC i2c_bus and i2c_dev
configuration values to string whether struct or yaml reads the
values as ints or strings. This allows for the configuration files
to use hex values for the I2C bus and device address without error.

Signed-off-by: Matt McKee <mmckee@phytec.com>