phytec_eeprom_flashtool: struct size, bit count, and api version fixups
This commit fixes an ill-defined struct size which was causing
additional data to be written to the EEPROM beyond the 32 bytes
we dedicate to the board information. Additionally, bit counts
were not being calculated correctly as numerical characters
were being converted to integers before counting bits. Lastly,
the PHYTEC EEPROM API version is bumped from zero to one.

Signed-off-by: Matt McKee <mmckee@phytec.com>