phytec_eeprom_flashtool: fix yaml int type for eeprom i2c address
Yaml has a defined integer type for tags. We should use it as specified.
A number will be interpreted as an int. If it has no prefix it will be
interpreted as decimal number. Entering 5e (without quotes) will be
interpreted as string. So currently the field uses mixed types and you
need to look at the code to understand how this works.

It is more intuitive to use the yaml data format as specified and
convert the output of the data to the needs of the use-case. With this
change the i2c_dev tag requires an int type and the output is converted
to a hex number string as required.

Signed-off-by: Stefan Müller-Klieser <s.mueller-klieser@phytec.de>