phytec_eeprom_flashtool: Add 'version' argument
Adds the -v/--version argument to print the current tool version.

Since the setup.py is not included in the package, the version is
now located in phytec_eeprom_flashtool/src/__init__.py and setup.py
will read this file to gain the package version. Therefore, the
version is only store at one location.

Signed-off-by: Daniel Schultz <d.schultz@phytec.de>