phytec_eeprom_flashtool: Read Product Name from Target
U-Boot SOM detection can now expose the product name and part
number to the Linux OS. They can be accessed with
  /proc/device-tree/phytec,...

Try to read the product name and disable the '-som' argument. That
allows to use 'phytec_eeprom_flashtool read' on our target without
added, for example, ' -som PCM-071' on a PCM-071 target.

Signed-off-by: Daniel Schultz <d.schultz@phytec.de>