Fix output when optiontree config is extended after data creation
This fix addresses the case when an eeprom binary is created and
afterwards the config is extended by new position(s).
In this case, the binary file has '\x00' at the new positions.
When the eeprom tool tries to display this information, an error occurs
as '\x00' can not be interpreted correctly.
This fix converts all '\x00' in the kit_opt string to "#" and interprets
it as "0" to get the corresponding option entry.

Signed-off-by: Marc Dzierzawa <M.Dzierzawa@phytec.de>